// types.ts
export enum Tab {
  ThinkingMode = 'Thinking Mode',
  FastResponder = 'Fast Responder',
  AccurateResponder = 'Accurate Responder',
  ImageStudio = 'Image Studio',
  ImageAnalyzer = 'Image Analyzer',
  VideoGenerator = 'Video Generation',
  VirtualDP = 'Virtual DP',
}

export type AspectRatio = '1:1' | '16:9' | '9:16' | '4:3' | '3:4';

export type VideoAspectRatio = '16:9' | '9:16';